fix(db): Analytics-Schema-Typdrift in frischen DBs reparieren + fresh-DB-Vertragstest
Die Baseline-Migration baute frische DBs mit falschen Spaltentypen (int4/real/text statt bigint/double/timestamptz) fuer die Analytics-Session-/Viewer-/Chat-/Raid-Tabellen. Die Live-DB ist korrekt, aber ein Neuaufbau aus den Rust-Migrationen haette die Kern-Analytics still gebrochen (Decode-/Operatorfehler).

Neue strikt geguardete Vorwaerts-Migration: No-op auf bereits korrekter DB (verifiziert: 0 DDL, kein relfilenode-Rewrite, kein FK-Diff), Korrektur nur bei abweichendem Typ; FK/PK/Sequenz sauber behandelt (FKs inkl. ON DELETE wiederhergestellt). Cast $6::int4->int8 fuer target_session_id (B4-024).

prod_contract um chat_messages/raid_retention erweitert. Neuer fresh_migrations_schema-Test + Runner-Skript (Wegwerf-TimescaleDB) beweist, dass die Migrationen aus dem Nichts das korrekte Schema bauen (rot ohne Fix, gruen mit) und faengt kuenftige Drift.

+++ rust/crates/tb-monitoring/src/raid_retention.rs
@@ -115,13 +115,13 @@ async fn compute_one(pool: &PgPool, raid: &RaidRow) -> Result<Outcome, sqlx::Err
     let new_to_target = count_new_to_target(pool, target_session_id, raid).await?;
     let new_chatters = count_new_chatters(pool, target_session_id, raid).await?;
 
-    // 5) Insert (target_session_id::int4-Cast), ON CONFLICT DO NOTHING.
+    // 5) Insert (target_session_id::int8-Cast), ON CONFLICT DO NOTHING.
     sqlx::query(
         "INSERT INTO twitch_raid_retention \
          (raid_id, from_broadcaster_login, to_broadcaster_login, viewer_count_sent, \
           executed_at, target_session_id, chatters_at_plus5m, chatters_at_plus15m, \
           chatters_at_plus30m, known_from_raider, new_to_target, new_chatters) \
-         VALUES ($1, $2, $3, $4, $5, $6::int4, $7, $8, $9, $10, $11, $12) \
+         VALUES ($1, $2, $3, $4, $5, $6::int8, $7, $8, $9, $10, $11, $12) \
          ON CONFLICT (raid_id, executed_at) DO NOTHING",
     )
     .bind(raid.id)
